Use 16bit size for external allocation, to limit the memory to only what can be addressed
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@19 96dfbd92-d197-e211-9ca9-00110a534b34

+++ v6502/mem.c
@@ -11,7 +11,7 @@
 #include "mem.h"
 #include "core.h"
 
-v6502_memory *v6502_createMemory(size_t size) {
+v6502_memory *v6502_createMemory(uint16_t size) {
 	// Allocate Memory Struct
 	v6502_memory *memory = malloc(sizeof(v6502_memory));
 	if (!memory) {
